requirements.txt
aiogram
python-dotenv
asyncpg
greenlet
gspread
apscheduler
babel